pyproject.toml
[project]
name = "voyager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "chardet>=5.2.0",
    "gymnasium>=1.1.1",
    "javascript>=1!1.2.2",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-qwq>=0.2.0",
    "langchain-chroma>=0.1.1",
    "minecraft-launcher-lib>=6.5",
    "openai>=1.93.0",
    "protobuf>=3.20.3,<4",
    "psutil>=7.0.0",
    "requests>=2.32.4",
    "setuptools>=80.9.0",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "python-dotenv>=1.0.1",
    "chromadb>=1.0.13",
    "posthog==6.0.0",
]
